Close the ssl file when done with it
The parser has two constructors.  One takes a stream and passes it to
the scanner.  The other takes a filename, opens it, and passes the
stream to the scanner.  Opening a file in the parser (esp. in the
constructor) isn't the best idea since the parser class would optionally
own the file and would need to close it at some point.

Move ownership of the file to the readSSLFile method (which has
responsibility of closing it).  Delete the parser constructor that takes
a filename.  Now there is no longer a source for the filename (for error
messages) so delete that as well; use "ssl" as the filename.  It could
take the filename as another parameter, but I will leave that for a
future improvement.