Resource cleanup
Merge BinaryFile::UnLoad() into the destructor.  Don't reuse the object,
create a new one instead.  Remove Prog::clear() for the same reason.

A Prog object will take care of deleting the BinaryFile and FrontEnd
when it is destructed so long as setFrontEnd has been called.  Update
all the tests to use 'delete prog;' where possible.

BinaryFileFactory is in charge of loading and unloading the library, but
it shouldn't be unloaded until all objects of that type have been
destroyed.  It makes more sense to unload the library whenever the
BinaryFile object is destroyed.  Remove the BinaryFileFactory class and
merge it into BinaryFile as the open() and close() static methods.

Remove unneeded linking of modules to libBinaryFile.la.