Tweak the configure defaults
Enable silent rules by default, so that the output is less cluttered and
easier to pick out error messages.  Can be disabled if I need to debug
the command line options.

Disable modules for now; I can get better backtrace info out of a core
dump if it's built as a monolithic executable.

Disable XML loading/saving.  Saving works (in fact it doesn't even need
expat) but loading is currently broken.  I think it's trying to create a
FrontEnd for it (is one even needed when the info is in XML form?)

Disable the garbage collector by default.  Seems to be causing stability
problems when running the GUI.