Adjust seek placement
Place the seek immediately before the read where possible.

Shouldn't need to check the stream state after a seek if it's followed
by a read that is also checked.  A bad seek should set one of the stream
failure bits, which will cause the read to be a no-op, then we check
those stream failure bits after the read and determine that the
operation failed.