Magic detection fixes
Move COFF detection higher, to respect the comment that HpSom is last
because it's "not particularly good magic".  HpSom was originally last,
then COFF was appended to the end when it was added.

Fix final byte check of HpSom to match the verification in
HpSomBinaryFile.cpp.  Should be checking buf[3], not buf[4].