Add manual output modification steps to mltk.sh
Adding the 'static' keyword to string tables doesn't appear to be needed
anymore with the "August 1st" version of the toolkit.  The binary
package available at SourceForge [1] must be using an older version.
Also there are some differences in where it puts line breaks.  The next
commit will regenerate the files with updated formatting.

[1] http://prdownloads.sourceforge.net/boomerang/sml-nj-with-toolkit-x86-linux.tar.gz?download