Regenerate files with updated line numbers
Uses mltk.sh in accordance with the Makefile rules, but minus the lines
that change nullptr and const.  Those will come in the next commit.

I'm not sure that the line numbers are correct, even after we modify
them in mltk.sh.  I can find several instances that are off-by-one.