Add original noweb sources for SLED files
pentium.nw and ppc.nw are the latest from version 0.5a of the toolkit.

An earlier version of mips.nw on the toolkit website generates an exact
match with the various mips files (minus a couple of path differences,
corrected in this commit).  Note that this is NOT the latest mips.nw
from version 0.5a of the toolkit.

The sparc files were likely generated from something even earlier than
the earlier version of sparc.nw (alongside that eariler version of
mips.nw).  I've copied the earlier version here, so note that this is
also NOT the latest sparc.nw from version 0.5a of the toolkit.  The line
directive differences in synth.spec are a clue.