Matcher cleanup
Remove brackets in matcher arms that don't appear to be necessary:  If
we don't need to restrict the scope of variables to just the matching
arm, then they're extra and can be removed.  Or if we do, but we're also
binding a name to one of the free variables, then it should create
brackets for us.  Otherwise it doesn't appear to create them, but
probably should for sanity.

Since 'else' at the start of a line is reserved when between the
match/endmatch pairs, clean up the remaining workarounds where it was
previously 'if (!exp)' instead of 'else'.  A close-bracket from the
preceding 'if', placed on the same line just before the 'else', is
enough to prevent the toolkit from interpreting it as an arm of the
matcher.

Try to balance all brackets around 'else' keywords.