Fix associativity of optables and casts
The parser needs to compare the precedence of what is on the stack vs.
the next terminal, so we have to assign a precedence to the '{' that
begins a cast and the NAME_LOOKUP that begins an optable.

No longer have a need to distinguish an exp_term from an exp.  This
works against the associativity/precedence declarations.