Avoid inverting some address calculations
I don't believe %pc in the SSL is used.  If it were, it should be
pushing nextPC as the return address, then branching to relocd.
As previously written, only one of those could be true at a time,
depending on whether %pc increments before or after the SSL statements.
Incrementing after doesn't really work either, unless relocd includes a
correction factor.

Now it's at least consistent in assuming %pc is incremented before, but
that may not be correct.