Exp printing fixes
Fix the printr() documentation to make it obvious that this should be
used for recursion in the default case.  Use printr() in a few cases
where the extra parens may be needed (opArrayIndex and opMemberAccess),
although the left side of those is not likely to be Binary or Ternary.

Print opPostVar with the other postfix unary operators.  This operator
may or may not serve the same purpose as opInitValueOf.

Ternary::printr() was throwing away the html parameter and doing the
default-case printing in a roundabout way (through operator <<).