Replace the Log class with std::ostream
Removes the limit that only types known to the Log class can be printed.

Printing of ADDRESSes needs to be cleaned up.  It would be ideal to have
std::ostream << ADDRESS implemented to print them in hex, but it's going
to require more work.  Probaby a new class to replace ADDRESS.