Clean up Type compatibility checks
FuncType vs. SizeType is not symmetric; make it so.  Use SizeType's
interpretation that they are never compatible.  I'm assuming the size
check in FuncType was really meant for comparing pointer-to-func.

For FuncType, comparing *this == *other is redundant with the signature
comparison as long as both have signatures, which is being enforced by
assert().

For NamedType, comparing *this == *other is also redundant with the name
comparison and adds nothing new.