Avoid BasicBlock::m_iNumOutEdges
From what I can tell, m_iNumOutEdges is supposed to be the final number
of outedges and may be larger than m_OutEdges.size() until all are
filled in.  The getNumOutEdges() and getOutEdge() interfaces make any
missing edges appear as null.

Replace with m_OutEdges.size() when nearby code is directly accessing
m_OutEdges[].  Note that we may not be iterating as many times as
intended (can't tell the original intent), but it'll avoid walking off
the end of the m_OutEdges vector.