Rewrite Cfg::splitForBranch() using Cfg::splitBB()
splitBB() adds a label to each bottom BB that it creates.
splitForBranch() was not adding labels, and I don't yet understand the
purpose of these labels, so remove them to match the previous behavior.

The out-edge vectors have the same order of elements as before.
The in-edge vector of B will have edges from S and R in reverse order
than before.  I don't think this should matter but it seems to affect
the DFG somehow.