Try to represent tables and ASCII art in doxygen
The Markdown tables are entirely insufficient, and I'm not going to
resort to HTML.  Wrap them in \code ... \endcode instead so they show up
as preformatted text.  Same for ASCII art, since indenting by 4 spaces
doesn't always work.  Sometimes it shows line numbers and sometimes it
doesn't.  Auto-linking is useful but other syntax highlighting is not.
Not sure if there's a better option.