Always allocate incomplete BBs
If Cfg::label() doesn't immediately split an existing BB, then the map
contains a null pointer that we have to account for.  This results in
basically two classes of incomplete BB, one where we've allocated the
container and can add edges to, another where we can't.  But in many
cases we don't account for the null pointer and dereference it anyway.
So always allocate the container and always assume it exists.