Clean up Exp::polySimplify() recursion
bMod should never be set to false in the callee, only in the caller.
The callee can change it to true.

TypedExp::polySimplify() now does a post-visit to match the others.

To speed it up, it would be useful to skip processing a node until its
children have been fully simplified.  RefExp::polySimplify() used to do
this by returning early if its child had changed.  However this results
in more polySimplify() iterations and repeating expensive checks on
already simplified nodes.  We need a more general solution.